Import 2.5.3

+++ sources/android/support/v4/widget/CircleImageView.java
@@ -36,7 +36,7 @@ class CircleImageView extends ImageView {
         } else {
             ShapeDrawable circle2 = new ShapeDrawable(new OvalShadow(this.mShadowRadius));
             setLayerType(1, circle2.getPaint());
-            circle2.getPaint().setShadowLayer((float) this.mShadowRadius, (float) shadowXOffset, (float) shadowYOffset, KEY_SHADOW_COLOR);
+            circle2.getPaint().setShadowLayer((float) this.mShadowRadius, (float) shadowXOffset, (float) shadowYOffset, (int) KEY_SHADOW_COLOR);
             int padding = this.mShadowRadius;
             setPadding(padding, padding, padding, padding);
             circle = circle2;
